# src/static/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Dashboard
  description: Endpoints for conveying dashboard data from backend to frontend.

servers:
- url: https://api.dashboard.deeplink.network

paths:
  /matrix:
    get:
      summary: Retrieve a segment of the matrix
      description: This API endpoint returns a segment of a matrix based on the provided row and column indexes.
      parameters:
        - in: query
          name: x
          schema:
            type: integer
          description: The starting row index.
        - in: query
          name: y
          schema:
            type: integer
          description: The ending row index.
        - in: query
          name: i
          schema:
            type: integer
          description: The starting column index.
        - in: query
          name: j
          schema:
            type: integer
          description: The ending column index.
        - in: query
          name: sort_by
          schema:
            type: string
          description: The metric to sort the matrix by. One of "liquidity", "price", or "volume"
      responses:
        '200':
          description: A matrix segment successfully returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/matrixSegment'
        '503':
          description: Indicates that the matrix has not been created yet, should only occur within the first 5 minutes of booting.

  /matrix_filter:
    get:
      summary: Filter the matrix by asset ID
      description: This API endpoint filters the matrix by the provided asset ID and returns the corresponding matrix segment.
      parameters:
        - in: query
          name: asset_id
          schema:
            type: string
          description: The asset ID to filter by.
      responses:
        '200':
          description: A filtered matrix segment successfully returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/matrixSegment'
components:
  schemas:
    matrixSegment:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
          description: The index of the columns.
        index:
          type: array
          items:
            type: string
          description: The index of the rows.
        data:
          type: array
          items:
            type: array
            items:
              type: object
              properties:
                pair:
                  type: object
                  description: The two assets in the cell expressed as SYMBOL_ID.
                  properties:
                    first:
                      type: object
                      properties:
                        id: 
                          type: string
                          description: The ID of the first asset in the pair.
                        symbol:
                          type: string
                          description: The symbol of the first asset in the pair.
                        name:
                          type: string
                          description: The name of the first asset in the pair.
                    second:
                      type: object
                      properties:
                        id:
                          type: string
                          description: The ID of the second asset in the pair.
                        symbol:
                          type: string
                          description: The symbol of the second asset in the pair.
                        name:
                          type: string
                          description: The name of the second asset in the pair.
                liquidity:
                  type: number
                  nullable: true
                  description: The total liquidity in USD of the pair across the supported DEXs.
                average_price:
                  type: number
                  nullable: true
                  description: The average price paid in USD when swapping the row asset for the column asset.
                price_movement_5m:
                  type: number
                  nullable: true
                  description: The percentage price change of the pair over the last 5 minutes.
                price_movement_1h:
                  type: number
                  nullable: true
                  description: The percentage price change of the pair over the last hour.
                price_movement_24h:
                  type: number
                  nullable: true
                  description: The percentage price change of the pair over the last 24 hours.
                volume_24h:
                  type: number
                  nullable: true
                  description: The volume in USD of the pair over the last 24 hours.
                safety_score:
                  type: number
                  nullable: true
                  description: A safety rating metric for the pair.
                exchanges:
                  type: array
                  items:
                    type: string
                  nullable: true
                  description: A list of the DEXs this pair is available on.
                diagonal:
                  type: boolean
                  description: A diagonal matrix item can be considered as irrelevant data.